requirements.txt
astroid==2.4.2
click==7.1.2
colorama==0.4.3
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.1.2
pylint==2.5.3
pypng==0.0.20
PyQRCode==1.2.1
pyzbar==0.1.8
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
